chore(@toss/react): refac useVisibilityEvent hook
