+++ packages/react/react/src/hooks/index.ts
@@ -26,4 +26,4 @@ export * from './useThrottle';
 export * from './useTimeout';
 export * from './useTimeoutQueue';
 export * from './useToggleState';
-export * from './useVisibilityEvent';
+export * from './useDocumentVisibilityChange';
