+++ packages/react/react/src/hooks/useDocumentVisibilityChange.en.md
@@ -1,0 +1,15 @@
+# useDocumentVisibilityChange
+
+Call the callback function when the document's visibilitychange event occurs and pass the visibility of document in parameter
+
+## Example
+
+```tsx
+function MyComponent() {
+  const visibleCallbackFunc = (isVisible: boolean) => {
+    alert('isVisible', isVisible);
+  };
+
+  useDocumentVisibilityChange(visibleCallbackFunc);
+}
+```
