--- packages/react/react/src/hooks/useVisibilityEvent.en.md
@@ -1,17 +1,0 @@
-# useVisibilityEvent
-
-Call the callback function when the document's visibilitychange event occurs.
-
-## Example
-
-```tsx
-type VisibilityState = Document['visibilityState'];
-
-function MyComponent() {
-  const visibleCallbackFunc = (visibilityState: VisibilityState) => {
-    alert(visibilityState);
-  };
-
-  useVisibilityEvent(visibleCallbackFunc);
-}
-```
